feat: add `fs.rs`.
Module that will contain utility methods to traverse and read the SAI file
system.

For now it is a WIP. This is mostly to test if the decrypting process works.
This is pretty much code that `Wunkolo` ( author of libsai ) already implement
it on their library.

refactor: make `Inode` part of the public API.