feat!: `SaiFileSystem` struct.
A `SaiFileSystem` encapsulates the logic of how the `SaiBlock`s within the SAI
file are structured. So, it will implement, how to traverse a SAI file system (
`SaiFsTraverser` trait, old `FileSystemVisitor` ), and how to read a particular
block within the file system.

The API is still mostly WIP ( I have a lot more ideas to implement ), but this
should be enough to make the ( future ) `SaiDocument` work.

BREAKING CHANGE:
- Replace `FileSystemVisitor` with `SaiFileSystem + SaiFsTraverser trait`.
- visitor -> traverser mod.