feat: SaiDocument::canvas().
Gets some canvas related information. i.e: alignment, width, height, etc.

fix: issue 6 on libsai?

I didn't still test it, but AFAIK this is the problem that the library was
facing while reading `layers`, and `sublayers`.

That will be the next `feat` introduced, so I will be able to verify if it is
actually works.

feat(reader): `next_stream_header()`.

_SAI stream are rather easy to read, but kinda annoying to work with._