refactor: remove `SaiBlock` and `block::Error`.
Now that I have a better understanding how I wan't to organize the API, the
`SaiBlock` trait wasn't that useful, because I never actually used the
`checksum()` method outside the `mod`.

Also removed `block::Error`, because this is a private API, so I expect to have
`good` input when calling their methods, and I can just `debug_assert()` in
case there is something wrong high in the chain ( shouldn't be able to happen ).