refactor: split `document` into different mods.
800 lines of code is way to much, so there will be a module for each
`SaiDocument` API call.

feat: InodeReader::{read_as_num(), read_with_size()}.

Decided to create a safe API for `read_as()` to read `numbers`. I didn't really
wanted to add `num-traits` as part of the public API, but I suppose that as long
they don't do a lot of breaking changes ( I doubt it ) it should be fine.

read_with_size() is pretty situational, but I guess I doesn't hurt to add it.

refactor: layer API was heavely changed.

There are still some things that I need to address, but the code now should be a
little more readable.

documentation: add more safety comments to unsafe blocks.

One of the reasons to of `read_as_num()` is to reduce the number of `unsafe`
calls through the code. That will focus the things that are really problematic
if any guarantees would not be meet.