feat: implement `Display` for `SaiDocument`.
The display trait of `SaiDocument` will print a `TreeView` of the underlying
layers inside it. This feature is behind the feature ( lol ) `tree_view`. As
always the feature is included as default.

fix: Docs-tests.

For some reason they werent running before. Now they do; I fixed the problems
with them to make them at least compile.