refactor!: rename all instances of `type` -> `kind`.
Since `type` is a reserved keyword on rust, you need to "escape" every variable
with that name. When I first wrote the library I didn't really tried to find a
good enough synonym, because I was OK with it? Anyways, this will make the API a
little less weird to read, I suppose ( at least for me it does ).

BREAKING CHANGE: InodeType -> InodeKind and LayerType -> LayerKind.