refactor: improve layer decompression algorithm.
The new code yields a 13-18% extra performance, mainly because
auto-vectorization of the `BGRA -> RGBA` inner loop.

This could probably be optimized further ( unsafe code most likely ), but for
now this is completely fine.

This also brings `itertools` dependency once again. I decided that as long a
dependency is small, it would be probably fine to add it to the project. I
will still try to remove as much of them as possible to not make compilation
times too big.