refactor: include test's sample bytes with `include_bytes`.
I actually worked on this a long time ago (probably like 8 months ago), but I
never push it from my laptop (oops).

The main motivation of this was to be able to run miri to check for UB. If you
check the diff you can see that the program had UB right from the beginning...

I'm gonna fix it on a follow up commit, because it would be worth to revisit
some of these old design decisions. I'm also gonna release a new version soon
after it, because I think it is about time.