refactor: overhaul the `Table` and `Data` Block impls.
This would be the start of the major overhaul that I have plan for the library.
Decided to start with the `block` api, because of the 2 problematic UBs that it
had.

The library _should_ be quite a bit faster (see reader.rs diff), but since I
can't run the previous lib version (because UB) I can't really benchmark the
results :(.

refactor: TableEntry fields are private.

refactor: various symbols name changes.

structs

- Inode       -> FatEntry
- InodeKind   -> FatKind
- InodeReader -> FatEntryReader

constants

- SAI_BLOCK_SIZE  -> BLOCK_SIZE
- BLOCKS_PER_PAGE -> BLOCKS_PER_SECTION

fields/methods

TableEntry

- ::next_block -> ::idx_of_next_block