refactor(reader): remove `read_as()` and `read_as_num()`.
Both of these methods were reporting UB through miri, so the best course of
action was to just remove them and implement safer methods.

The alternative of read_as() is read_array() (that was the most meaningful usage
of it, other than in the read_as_num() impl), and read_as_num() was replaced
with read_{u8, u16, u32, i32, u64}() methods.

Performance is about the same between both implementations.