feat(table): implement `Debug` for `IntoIter`.
I forgot to add this back when I implement it, hopefully now I have more time to
finish refactoring this crate once and for all.

refactor: minor changes to `block` module.

I think that final design for the block api is almost done; there are a couple
of things that I need some more polish, but I should be able to make the module
public soon.