revert: all mask changes.
This changes slowed down the program about 35% to 80%. I imagine that it is
probably, because LLVM can not longer auto-vectorize the important loops.

There is still a 2% to 5% perfomance degradation with the changes merged from
`main`; if I have to guess, the new block API might be the problem; it could
also be the new `sort_layers` method, but I doubt it.