refactor: new and improved `cipher` API.
Yet another refactor for the `cipher` API; some changes on `feat/vfs` needed
more flexibility than the old API provided, so might as well fix those issues
and redesign the whole thing.

I have some left to do on regards to this branch (like writing docs), but this
should be enough to start experimenting with `feat/vfs`.

---

For now, I gonna keep this out of main for 2 reasons:

1) This relies on `adt_const_params` which is still unstable. I would propose to
upstream if they are willing to allow arrays on stable as a minimal extension
while waiting the whole feature to get stabilize.

2) This changes are mainly concerned about the new `vfs` API (`feat/vfs`). That
means that there is not really a lot of interest trying to make this work with
the current API, because I can't merge this to main regardless, and there isn't
any performance wins that would justify changing the existent code (even though
the changes would be _mostly_ "automatic").