refactor!: move `utils::pixel_ops` to `crate::pixel_ops`.
Was considering deprecating the module, to make it private in a later version,
but currently it is not clear if that is the best solution, so for now it stays.

refactor!: `premultiplied_to_straight` now panics if `pixels % 4 != 0`.

refactor: rename `utils` to `internals`.

Was considering gating the visibility of the module with a feature flag. There
are currently no out-of-tree crates that would need such functionality, so I
don't think there is a lot of value doing it.