feat: initial file system implementation.
I have been working with this out-of-tree for quite a while. Now that I have
something that works I want to use it with the old API. This also keeps the
code from `main` around in case I need to do some benchmarks or test some
regression.

Once this branch is ready, I'm gonna rename all `*_` modules and remove the
modules with the old implementations.